Split squashfuse_ll into a lib and executable
This makes it easier for people to statically link squashfuse_ll into
their application, which can in turn be used to more easily ship
self-exectracting binaries (e.g. facebook xar or appimage)
