+++ ll.h
@@ -71,4 +71,95 @@ sqfs_err sqfs_ll_iget(fuse_req_t req, sqfs_ll_i *lli, fuse_ino_t i);
 
 int sqfs_ll_daemonize(int fg);
 
+void sqfs_ll_op_getattr(fuse_req_t req, fuse_ino_t ino,
+	struct fuse_file_info *fi);
+
+void sqfs_ll_op_opendir(fuse_req_t req, fuse_ino_t ino,
+	struct fuse_file_info *fi);
+
+void sqfs_ll_op_create(fuse_req_t req, fuse_ino_t parent, const char *name,
+	mode_t mode, struct fuse_file_info *fi);
+
+void sqfs_ll_op_releasedir(fuse_req_t req, fuse_ino_t ino,
+	struct fuse_file_info *fi);
+
+size_t sqfs_ll_add_direntry(fuse_req_t req, char *buf, size_t bufsize,
+	const char *name, const struct stat *st, off_t off);
+
+void sqfs_ll_op_readdir(fuse_req_t req, fuse_ino_t ino, size_t size,
+	off_t off, struct fuse_file_info *fi);
+
+void sqfs_ll_op_lookup(fuse_req_t req, fuse_ino_t parent,
+		const char *name);
+
+void sqfs_ll_op_open(fuse_req_t req, fuse_ino_t ino,
+		struct fuse_file_info *fi);
+
+void sqfs_ll_op_release(fuse_req_t req, fuse_ino_t ino,
+		struct fuse_file_info *fi);
+
+void sqfs_ll_op_read(fuse_req_t req, fuse_ino_t ino,
+		size_t size, off_t off, struct fuse_file_info *fi);
+
+void sqfs_ll_op_readlink(fuse_req_t req, fuse_ino_t ino);
+
+void sqfs_ll_op_listxattr(fuse_req_t req, fuse_ino_t ino, size_t size);
+
+void sqfs_ll_op_getxattr(fuse_req_t req, fuse_ino_t ino,
+		const char *name, size_t size
+#ifdef FUSE_XATTR_POSITION
+		, uint32_t position
+#endif
+		);
+
+void sqfs_ll_op_forget(fuse_req_t req, fuse_ino_t ino,
+		unsigned long nlookup);
+
+void stfs_ll_op_statfs(fuse_req_t req, fuse_ino_t ino);
+
+
+/* Helpers to abstract out FUSE 2.5 vs 3.0+ differences */
+
+typedef struct {
+	int fd;
+	struct fuse_session *session;
+#if FUSE_USE_VERSION < 30
+	struct fuse_chan *ch;
+#endif
+} sqfs_ll_chan;
+
+#if FUSE_USE_VERSION >= 30
+sqfs_err sqfs_ll_mount(
+		sqfs_ll_chan *ch,
+		const char *mountpoint,
+		struct fuse_args *args,
+        struct fuse_lowlevel_ops *ops,
+        size_t ops_size,
+        void *userdata);
+
+void sqfs_ll_unmount(sqfs_ll_chan *ch, const char *mountpoint);
+
+#else /* FUSE_USE_VERSION >= 30 */
+
+void sqfs_ll_unmount(sqfs_ll_chan *ch, const char *mountpoint);
+
+sqfs_err sqfs_ll_mount(
+		sqfs_ll_chan *ch,
+		const char *mountpoint,
+		struct fuse_args *args,
+		struct fuse_lowlevel_ops *ops,
+		size_t ops_size,
+		void *userdata);
+
+#endif /* FUSE_USE_VERSION >= 30 */
+
+void alarm_tick(int sig);
+
+void setup_idle_timeout(struct fuse_session *se, unsigned int timeout_secs);
+
+void teardown_idle_timeout();
+
+sqfs_ll *sqfs_ll_open(const char *path, size_t offset);
+
+
 #endif
